requirements.txt
Flask>=2.0.0
Flask-SocketIO==5.3.6
python-telegram-bot>=13.7
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
pandas>=1.3.0
APScheduler>=3.8.1
gevent==23.9.1
gevent-websocket==0.10.1
redis>=4.0.2
Pillow>=8.2.0
requests>=2.26.0
plotly>=5.3.1
python-dotenv>=0.19.0
gunicorn==21.2.0
pytz==2024.1
eventlet==0.35.2
python-magic>=0.4.24
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.6b0
flake8>=3.9.2
mypy>=0.910
pyspellchecker>=0.6.2
python-language-tool>=2.5.3
celery>=5.1.2
SQLAlchemy>=1.4.23
alembic>=1.7.3
psycopg2-binary>=2.9.1
fastapi>=0.68.1
uvicorn>=0.15.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
